Latex lösningar + feedback
Lade till lösningsuppladdning m.h.a latex och en feedback meddelande system så användaren vet om sin lösning har laddats upp eller ej. Nu går det också att klistra in lösningar direkt utan att behöva ladda ner lösning

+++ backend/routes/solutions.js
@@ -30,8 +30,9 @@ router.post("/delete/:questionNum/:solutionID", async (req, res) => {
     const solutionInSolutionsArr = solutionDoc.solutions.find(x => x.solutionID === solutionID);
     if(!solutionInSolutionsArr) return sendMsg(res, "Not found.", 400);
     if(solutionInSolutionsArr.ID !== userID) return sendMsg(res, "Unauthorized.", 400);
-
-    await imgur.deleteImage(solutionInSolutionsArr.deletehash).catch(err => console.log(err));
+    if(solutionInSolutionsArr.deletehash) {
+        await imgur.deleteImage(solutionInSolutionsArr.deletehash).catch(err => console.log(err));
+    }
     solutionDoc.solutions = [...solutionDoc.solutions.filter(x => x.solutionID !== solutionID)];
     await solutionDoc.save().catch(err => console.log(err));
